@@ -99,7 +82,10 @@
     "@types/react": "^18.0.0",
     "@types/react-csv": "^1.1.8",
     "@types/react-dom": "^18.0.0",
-    "pino-pretty": "^10.2.0",
-    "typescript": "^5.0.0"
+    "@openstad-headless/eslint-config": "*",
+    "tailwind-merge": "^1.14.0",
+    "tailwindcss": "3.3.3",
+    "tailwindcss-animate": "^1.0.7",
+    "typescript": "5.3.3"
   }
 }
